requirements.txt
jira==1.0.3
oauthlib==1.0.3
requests==2.9.1
requests-oauthlib==0.6.0
requests-toolbelt==0.5.1
six==1.10.0
suds-jurko==0.6
tlslite==0.4.9
